requirements.txt
pyyaml
nextcord